requirements.txt
pyyaml
nextcord